Various utility functions

Utility functions, usually called within the package and not exported.

format_metrics_instant_data(x)

format_metrics_range_data(x)

rename_metrics_data_frame(x)

parse_timeout(timeout_val)

response_check(response)

metrics_check(metric)

clean_metric_names(metric)

Arguments

x
A data frame generated by Prometheus query

timeout_val
A timeout value assigned by user.

response
httr response object

metrics
Generated metrics object

metrics
Generated metrics object

Utility function for parsing metrics output

Formats metrics data from the current query

Utility function for parsing range metrics output

Formats metrics data passed by range query.

Utility function cleaning column names

As there is no clarity with respect to the returned columns the function:

Chaneges names of known columns like name

Removes special characters from all columns

Utility function for parsing timeout argument parse_timeout

Fixes potentially problematic timeout argument.

Checking Prometheus Server response

Returns more meanigful error code

Check Metrics Object

Check if metrics objects contains data. If there is no data for time period referenced in the query this function will return an error.

Clean metrics object column names

NA